added nullptr test for a component

+++ TaskHelper.cpp
@@ -81,7 +81,10 @@ void TaskHelper::cancel_task(EntitySystem& ents, std::size_t task_id)
 	if(comp)
 	{
 		auto handler = ents.get_component<TaskHandlerComponent>(comp->source);
-		handler->curr_task = Component::NO_ENTITY;
-		DestructorHelper::destroy(ents, task_id);
+		if(handler)
+		{
+			handler->curr_task = Component::NO_ENTITY;
+			DestructorHelper::destroy(ents, task_id);
+		}
 	}
 }
